fix(web): address open-code-review findings on the BHK explainer
- makeDm moved to the shared explainer-cities module (was duplicated in
  branch-bound-algo and bhk-algo); both import and re-export it
- computeRoute: guard against end === -1 (no complete tour) instead of
  crashing on pred[-1]
- bhk.tsx: replace the CityMap nested ternary with if/else, bounds-check
  selected.row, extract the shared masks array, type historyRef as SimState[]
- restore the bestTour closed-cycle fix (start re-appended) that was
  accidentally reverted while untangling branch history
